Organize plot modules and deleted deprecated ones
