+++ auxiliary_tools/cdat_regression_testing/655-zonal-mean-2d/655_zonal_mean_2d_stratosphere_run_script.py
@@ -1,6 +1,6 @@
 from auxiliary_tools.cdat_regression_testing.base_run_script import run_set
 
 SET_NAME = "zonal_mean_2d_stratosphere"
-SET_DIR = "655-zonal-mean-2d-stratosphere"
+SET_DIR = "655-zonal-mean-2d-stratosphere-debug"
 
 run_set(SET_NAME, SET_DIR)
