+++ e3sm_diags/plot/zonal_mean_2d_plot.py
